+++ lib/jq.rb
@@ -45,7 +45,7 @@ module JQ
   ##
   # The gem version number
   #
-  VERSION = "1.0.0"
+  VERSION = "1.1.0"
 
   ##
   # Base exception class for all jq-related errors.